refactor: mip map for textures
