+++ include/Engine/Command.h
@@ -1,19 +1,7 @@
 #pragma once
 #include "Device.h"
 #include "Sync.h"
-
-class CommandPool {
- private:
-  std::shared_ptr<Device> _device;
-  QueueType _type;
-  VkCommandPool _commandPool;
-
- public:
-  CommandPool(QueueType type, std::shared_ptr<Device> device);
-  VkCommandPool& getCommandPool();
-  QueueType getType();
-  ~CommandPool();
-};
+#include "Pool.h"
 
 class CommandBuffer {
  private:
