+++ src/Engine/Pool.cpp
@@ -1,0 +1,48 @@
+#include "Descriptor.h"
+
+CommandPool::CommandPool(QueueType type, std::shared_ptr<Device> device) {
+  _device = device;
+  _type = type;
+  VkCommandPoolCreateInfo poolInfo{};
+  poolInfo.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
+  poolInfo.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
+  poolInfo.queueFamilyIndex = device->getSupportedFamilyIndex(type).value();
+
+  if (vkCreateCommandPool(device->getLogicalDevice(), &poolInfo, nullptr, &_commandPool) != VK_SUCCESS) {
+    throw std::runtime_error("failed to create graphics command pool!");
+  }
+}
+
+QueueType CommandPool::getType() { return _type; }
+
+VkCommandPool& CommandPool::getCommandPool() { return _commandPool; }
+
+CommandPool::~CommandPool() { vkDestroyCommandPool(_device->getLogicalDevice(), _commandPool, nullptr); }
+
+DescriptorPool::DescriptorPool(int number, std::shared_ptr<Device> device) {
+  _device = device;
+
+  std::array<VkDescriptorPoolSize, 4> poolSizes{};
+  poolSizes[0].type = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
+  poolSizes[0].descriptorCount = static_cast<uint32_t>(number);
+  poolSizes[1].type = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
+  poolSizes[1].descriptorCount = static_cast<uint32_t>(number);
+  poolSizes[2].type = VK_DESCRIPTOR_TYPE_STORAGE_IMAGE;
+  poolSizes[2].descriptorCount = static_cast<uint32_t>(number);
+  poolSizes[3].type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
+  poolSizes[3].descriptorCount = static_cast<uint32_t>(number);
+
+  VkDescriptorPoolCreateInfo poolInfo{};
+  poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
+  poolInfo.poolSizeCount = static_cast<uint32_t>(poolSizes.size());
+  poolInfo.pPoolSizes = poolSizes.data();
+  poolInfo.maxSets = static_cast<uint32_t>(number);
+
+  if (vkCreateDescriptorPool(device->getLogicalDevice(), &poolInfo, nullptr, &_descriptorPool) != VK_SUCCESS) {
+    throw std::runtime_error("failed to create descriptor pool!");
+  }
+}
+
+VkDescriptorPool& DescriptorPool::getDescriptorPool() { return _descriptorPool; }
+
+DescriptorPool::~DescriptorPool() { vkDestroyDescriptorPool(_device->getLogicalDevice(), _descriptorPool, nullptr); }
