+++ src/Graphic/LightManager.cpp
@@ -97,10 +97,9 @@ LightManager::LightManager(std::shared_ptr<CommandBuffer> commandBufferTransfer,
 
   // stub texture
   _stubTexture = std::make_shared<Texture>("../data/Texture1x1.png", _state->getSettings()->getLoadTextureColorFormat(),
-                                           VK_SAMPLER_ADDRESS_MODE_REPEAT, 1, commandBufferTransfer,
-                                           _state->getSettings(), _state->getDevice());
+                                           VK_SAMPLER_ADDRESS_MODE_REPEAT, 1, commandBufferTransfer, _state);
   _stubCubemap = std::make_shared<Cubemap>(
-      std::vector<std::string>(6, "../data/Texture1x1.png"), _state->getSettings()->getLoadTextureColorFormat(),
+      std::vector<std::string>(6, "../data/Texture1x1.png"), _state->getSettings()->getLoadTextureColorFormat(), 1,
       VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL, VK_IMAGE_ASPECT_COLOR_BIT,
       VK_IMAGE_USAGE_TRANSFER_DST_BIT | VK_IMAGE_USAGE_SAMPLED_BIT, commandBufferTransfer, _state);
 
@@ -377,7 +376,7 @@ std::shared_ptr<PointLight> LightManager::createPointLight(std::tuple<int, int> 
   std::vector<std::shared_ptr<Cubemap>> depthCubemap;
   for (int i = 0; i < _state->getSettings()->getMaxFramesInFlight(); i++) {
     depthCubemap.push_back(std::make_shared<Cubemap>(
-        resolution, _state->getSettings()->getDepthFormat(), VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL,
+        resolution, _state->getSettings()->getDepthFormat(), 1, VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL,
         VK_IMAGE_ASPECT_DEPTH_BIT, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT | VK_IMAGE_USAGE_SAMPLED_BIT,
         _commandBufferTransfer, _state));
   }
@@ -411,8 +410,7 @@ std::shared_ptr<DirectionalLight> LightManager::createDirectionalLight(std::tupl
                         VK_IMAGE_ASPECT_DEPTH_BIT, 1, 1, _commandBufferTransfer);
     auto imageView = std::make_shared<ImageView>(image, VK_IMAGE_VIEW_TYPE_2D, 1, 0, 1, VK_IMAGE_ASPECT_DEPTH_BIT,
                                                  _state->getDevice());
-    depthTexture.push_back(
-        std::make_shared<Texture>(VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_BORDER, imageView, _state->getDevice()));
+    depthTexture.push_back(std::make_shared<Texture>(VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_BORDER, 1, imageView, _state));
   }
 
   auto light = std::make_shared<DirectionalLight>();
